Extract rowset-oriented tests in a separate file too
No real changes, just move the tests in their own file.
